Added GDT, etc.

+++ include/brunix/defs.h
@@ -1,6 +1,6 @@
 /**
  * @author Bruno Masci
- * @brief Some common definitions
+ * @brief Some common kernel definitions
  */
 
 #ifndef __DEFS_H__
@@ -21,7 +21,7 @@
  * NOTE: It only works for 32-bits numbers!
  * TODO CORREGIR Si no es exaxtamente multiplo de 2
  */
-static inline long roundup_binary(uint32_t n) {
+static inline uint32_t roundup_binary(uint32_t n) {
     n--;
     n |= n >> 1;
     n |= n >> 2;
